virt-controller-vmi: separate storage functions
This move storage related functions and event handlers to storage.go
