+++ pkg/virt-controller/watch/vmi/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "datavolumes.go",
+        "storage.go",
         "vmi.go",
         "volume-hotplug.go",
     ],
